Replace old file with refactored file
