+++ package.json
@@ -13,7 +13,7 @@
     "type": "git",
     "url": "https://github.com/LBBO/node-netflix2.git"
   },
-  "main": "./lib/netflix2.class.js",
+  "main": "lib/netflix2.js",
   "dependencies": {
     "cheerio": "^0.20.0",
     "extend": "^3.0.0",
